refactor: 의존성 관리 컨테이너 패키지 분리
