+++ di/src/main/java/woowacourse/shopping/di/ClassQualifier.kt
@@ -1,5 +1,8 @@
 package woowacourse.shopping.di
 
+import woowacourse.shopping.di.container.AnnotationQualifier
+import woowacourse.shopping.di.container.DependencyType
+
 data class ClassQualifier(
     val dependencyType: DependencyType,
     val annotationQualifier: AnnotationQualifier,
