+++ di/src/main/java/woowacourse/shopping/di/DependencyProvider.kt
@@ -1,5 +1,8 @@
 package woowacourse.shopping.di
 
+import woowacourse.shopping.di.container.DependencyInstance
+import woowacourse.shopping.di.container.ImplementationClass
+
 interface DependencyProvider {
     fun <T : DependencyInstance> getInstance(): T?
