+++ di/src/test/java/woowacourse/shopping/di/ApplicationDependencyContainerTest.kt
@@ -3,6 +3,8 @@ package woowacourse.shopping.di
 import com.google.common.truth.Truth.assertThat
 import org.junit.Before
 import org.junit.Test
+import woowacourse.shopping.di.container.ApplicationDependencyContainer
+import woowacourse.shopping.di.container.DependencyContainer
 
 class ApplicationDependencyContainerTest {
     private lateinit var dependencyContainer: DependencyContainer
